requirements/base.txt
alembic==1.7.5
aniso8601==9.0.1
attrs==21.4.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
Mako==1.1.6
MarkupSafe==2.0.1
PyMySQL==1.0.2
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
Werkzeug==2.0.2
